apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fairmind-backend"
version = "0.1.0"
description = "AI Governance Platform Backend"
authors = [{name = "Fairmind Team", email = "support@fairmind.xyz"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi>=0.121.1",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.4",
    "python-multipart>=0.0.20",
    "PyJWT>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "websockets>=15.0.1",
    "pandas>=2.3.3",
    "numpy>=2.3.4",
    "supabase>=2.24.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "joblib>=1.5.2",
    "pydantic-settings>=2.11.0",
    "python-json-logger>=4.0.0",
    "databases[postgresql]>=0.9.0",
    "redis>=7.0.1",
    "psutil>=7.1.3",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "sentry-sdk[fastapi]>=2.43.0",
    "aiosmtplib>=5.0.0",
    "boto3>=1.40.69",
    "httpx>=0.28.1",
    "gunicorn>=23.0.0",
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "reportlab>=4.0.0",
    "aif360>=0.6.1",
    "tensorflow>=2.15.0",
    "duckdb>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pylint>=2.17.0",
    "autopep8>=2.0.0",
    "coverage>=7.3.0",
    "httpx>=0.25.0",
    "factory-boy>=3.3.0",
    "faker>=19.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.25.0",
    "factory-boy>=3.3.0",
    "faker>=19.0.0",
    "pytest-playwright>=0.4.0",
    "playwright>=1.40.0",
]

[project.urls]
Homepage = "https://fairmind.xyz"
Repository = "https://github.com/adhit-r/fairmind"
Documentation = "https://docs.fairmind.xyz"
"Bug Tracker" = "https://github.com/adhit-r/fairmind/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "config*", "database*", "services*", "utils*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["api", "config", "database", "services", "utils"]
known_third_party = ["fastapi", "pydantic", "sqlalchemy", "pandas", "numpy"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true
error_summary = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "numpy.*",
    "sklearn.*",
    "supabase.*",
    "psycopg2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=api",
    "--cov=config",
    "--cov=database",
    "--cov=services",
    "--cov=utils",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "--asyncio-mode=auto",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["api", "config", "database", "services", "utils"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", "env"]
skips = ["B101", "B601"]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
]

[tool.pylint.format]
max-line-length = "100"

[tool.pylint.design]
max-args = 8
max-locals = 20
max-returns = 6
max-branches = 15
max-statements = 60
